requirements.txt
tensorflow-gpu==1.13.2
tqdm==4.42.1
bert_serving_server==1.10.0
numpy==1.18.1
bert_serving==0.0.1
